+++ include/SdlSceneGraph2.h
@@ -38,7 +38,7 @@ namespace astu {
         // Inherited via VertexBuffer2Builder
         virtual VertexBuffer2Builder& AddVertex(float x, float y) override;
         virtual VertexBuffer2Builder& Reset() override;
-        virtual std::unique_ptr<VertexBuffer2> Build() override;        
+        virtual std::shared_ptr<VertexBuffer2> Build() override;        
 
     private:
         /** The vertices used for the buffer to build. */