+++ src/math/MathUtils.cpp
@@ -1,0 +1,20 @@
+/*
+ * ASTU - AST Utilities
+ * A collection of Utilities for Applied Software Techniques (AST).
+ * 
+ * Copyright (c) 2020, 2021 Roman Divotkey, Nora Loimayr. All rights reserved.
+ */
+
+#pragma once
+
+// Local includes
+#include "MathUtils.h"
+
+namespace astu {
+
+    const double MathUtils::PId = 3.14159265358979323846;
+    const double MathUtils::PI2d = PId * 2;
+    const float MathUtils::PIf = static_cast<float>(PId);
+    const float MathUtils::PI2f = static_cast<float>(PI2d);
+
+} // end of namespace